2014-01-22  Brent Fulgham  <bfulgham@apple.com>

Source/JavaScriptCore:

    [Win] Update project and solution files for 64-bit builds.
    https://bugs.webkit.org/show_bug.cgi?id=127457

    Reviewed by Eric Carlson.

    * JavaScriptCore.vcxproj/JavaScriptCore.submit.sln: Add 64-bit target.
    * JavaScriptCore.vcxproj/JavaScriptCore.vcxproj: Specify v100 target
    for all builds on this branch.
    * JavaScriptCore.vcxproj/JavaScriptCore.vcxproj.filters: Add missing
    file from project view.
    * JavaScriptCore.vcxproj/jsc/jsc.vcxproj: Specify v100 target for
    all builds on this branch.
    * JavaScriptCore.vcxproj/testRegExp/testRegExp.vcxproj: Ditto
    * JavaScriptCore.vcxproj/testapi/testapi.vcxproj: Ditto

Source/WebCore:

    [Win] Update project and solution files for 64-bit builds
    https://bugs.webkit.org/show_bug.cgi?id=127457

    Reviewed by Eric Carlson.

    * WebCore.vcxproj/QTMovieWin/QTMovieWin.vcxproj: Specify v100 target
    * WebCore.vcxproj/WebCore.submit.sln: Add x64 targets
    * WebCore.vcxproj/WebCore.vcxproj: Specify v100 target for all builds
    on this branch. Also exclude 32-bit specific assembly when building
    64-bit target, and correct bad merge from branch.
    * WebCore.vcxproj/WebCore.vcxproj.filters: Correct branch merge
    * config.h: Handle 64-bit type definitions.
    * platform/graphics/ca/win/PlatformCAAnimationWin.cpp:
    (PlatformCAAnimation::setFromValue): Use CGFloat to support 64-bit builds
    (PlatformCAAnimation::setToValue): Ditto
    (PlatformCAAnimation::setValues): Ditto
    * platform/graphics/win/FontCustomPlatformData.cpp:
    (WebCore::FontCustomPlatformData::fontPlatformData): Add cast to
    support 32- and 64-bit targets.
    * platform/graphics/win/GraphicsContextCGWin.cpp:
    (WebCore::GraphicsContext::drawLineForDocumentMarker): Use CGFloat to
    support 64-bit builds.
    * platform/win/PasteboardWin.cpp:
    (WebCore::Pasteboard::writeURLToDataObject): Specialize std::min to
    work on 32- and 64-bit code.
    (WebCore::createGlobalImageFileDescriptor): Ditto
    * platform/win/StructuredExceptionHandlerSuppressor.cpp:
    (WebCore::StructuredExceptionHandlerSuppressor::StructuredExceptionHandlerSuppressor):
    Comment out 32-bit inline assembly.
    (WebCore::StructuredExceptionHandlerSuppressor::~StructuredExceptionHandlerSuppressor):
    Ditto

Source/WebKit:

    [Win] Update project and solution files for 64-bit builds
    https://bugs.webkit.org/show_bug.cgi?id=127457

    Reviewed by Eric Carlson.

    * WebKit.vcxproj/Interfaces/Interfaces.vcxproj: Specify v100 target
    for all builds on this branch.
    * WebKit.vcxproj/WebKit.submit.sln: Add 64-bit target.
    * WebKit.vcxproj/WebKit/WebKit.vcxproj: Specify v100 target for all
    builds on this branch.
    * WebKit.vcxproj/WebKit/WebKitApple.props: Remove unused QTMovieWin
    library dependency.
    * WebKit.vcxproj/WebKitExportGenerator/WebKitExportGenerator.vcxproj:
    Specify v100 target for all builds on this branch.
    * WebKit.vcxproj/WebKitGUID/WebKitGUID.vcxproj: Ditto

Source/WTF:

    [Win] Update project and solution files for 64-bit builds.
    https://bugs.webkit.org/show_bug.cgi?id=127457

    Reviewed by Eric Carlson.

    * WTF.vcxproj/WTF.submit.sln: Add x64 targets.
    * WTF.vcxproj/WTF.vcxproj: Specify v100 target for all builds on this
    branch.



Canonical link: https://commits.webkit.org/135908.567@safari-537.75-branch
git-svn-id: https://svn.webkit.org/repository/webkit/branches/safari-537.75-branch@162624 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/ca/win/PlatformCAAnimationWin.cpp
@@ -365,7 +365,7 @@ void PlatformCAAnimation::setFromValue(const FloatPoint3D& value)
     if (animationType() != Basic)
         return;
 
-    float a[3] = { value.x(), value.y(), value.z() };
+    CGFloat a[3] = { value.x(), value.y(), value.z() };
     RetainPtr<CACFVectorRef> v = adoptCF(CACFVectorCreate(3, a));
     CACFAnimationSetFromValue(m_animation.get(), v.get());
 }
@@ -375,7 +375,7 @@ void PlatformCAAnimation::setFromValue(const WebCore::Color& value)
     if (animationType() != Basic)
         return;
 
-    float a[4] = { value.red(), value.green(), value.blue(), value.alpha() };
+    CGFloat a[4] = { value.red(), value.green(), value.blue(), value.alpha() };
     RetainPtr<CACFVectorRef> v = adoptCF(CACFVectorCreate(4, a));
     CACFAnimationSetFromValue(m_animation.get(), v.get());
 }
@@ -418,7 +418,7 @@ void PlatformCAAnimation::setToValue(const FloatPoint3D& value)
     if (animationType() != Basic)
         return;
 
-    float a[3] = { value.x(), value.y(), value.z() };
+    CGFloat a[3] = { value.x(), value.y(), value.z() };
     RetainPtr<CACFVectorRef> v = adoptCF(CACFVectorCreate(3, a));
     CACFAnimationSetToValue(m_animation.get(), v.get());
 }
@@ -428,7 +428,7 @@ void PlatformCAAnimation::setToValue(const WebCore::Color& value)
     if (animationType() != Basic)
         return;
 
-    float a[4] = { value.red(), value.green(), value.blue(), value.alpha() };
+    CGFloat a[4] = { value.red(), value.green(), value.blue(), value.alpha() };
     RetainPtr<CACFVectorRef> v = adoptCF(CACFVectorCreate(4, a));
     CACFAnimationSetToValue(m_animation.get(), v.get());
 }
@@ -485,7 +485,7 @@ void PlatformCAAnimation::setValues(const Vector<FloatPoint3D>& value)
         
     RetainPtr<CFMutableArrayRef> array = adoptCF(CFArrayCreateMutable(0, value.size(), &kCFTypeArrayCallBacks));
     for (size_t i = 0; i < value.size(); ++i) {
-        float a[3] = { value[i].x(), value[i].y(), value[i].z() };
+        CGFloat a[3] = { value[i].x(), value[i].y(), value[i].z() };
         RetainPtr<CACFVectorRef> v = adoptCF(CACFVectorCreate(3, a));
         CFArrayAppendValue(array.get(), v.get());
     }
@@ -500,7 +500,7 @@ void PlatformCAAnimation::setValues(const Vector<WebCore::Color>& value)
         
     RetainPtr<CFMutableArrayRef> array = adoptCF(CFArrayCreateMutable(0, value.size(), &kCFTypeArrayCallBacks));
     for (size_t i = 0; i < value.size(); ++i) {
-        float a[4] = { value[i].red(), value[i].green(), value[i].blue(), value[i].alpha() };
+        CGFloat a[4] = { value[i].red(), value[i].green(), value[i].blue(), value[i].alpha() };
         RetainPtr<CACFVectorRef> v = adoptCF(CACFVectorCreate(4, a));
         CFArrayAppendValue(array.get(), v.get());
     }
